閲覧専用ページをコンポーネントに分離
